+++ dashboard/client/components/SubMenu/index.tsx
@@ -5,30 +5,35 @@ import React, { Fragment } from 'react'
 import './styles.scss'
 
 const SubMenu = ({ spaces, space }) => {
-  const x = (id, name) => _ => Router.push({ pathname: '/home', query: { space: id } })
+  const x = (id, name) => _ => Router.push({ pathname: '/tickets', query: { space: id } })
 
   return (
     <div className="sub-menu">
       <div className="sub-menu__header">Developer Relations Dashboard</div>
       <div className="sub-menu__wrapper">
-        <div className={'sub-menu__title' + (!space ? '-selected' : '')}>
+        <div className={'sub-menu__title' + (space ? '' : '-selected')}>
           <div>All Spaces</div>
           <div>378</div>
         </div>
         {spaces &&
-          spaces.map(({ id, name }) => (
-            <Fragment key={id}>
-              <div
-                onClick={x(id, name)}
-                className={'sub-menu__items' + (space === id ? '-selected' : '')}
-                key={id}
-              >
-                <div>{name}</div>
-                <div>{Math.round(Math.random() * 100)}</div>
-              </div>
-              <div className="test" />
-            </Fragment>
-          ))}
+          spaces.map(({ id, name, totalCount }) => {
+            if (name === 'Default' || name === 'Help') {
+              return null
+            }
+            return (
+              <Fragment key={id}>
+                <div
+                  onClick={x(id, name)}
+                  className={'sub-menu__items' + (space === id ? '-selected' : '')}
+                  key={id}
+                >
+                  <div>{name}</div>
+                  <div>{totalCount}</div>
+                </div>
+                <div className="test" />
+              </Fragment>
+            )
+          })}
       </div>
       <div className="sub-menu__background" />
     </div>
