+++ dashboard/client/components/SubMenu/styles.scss
@@ -27,7 +27,7 @@
   top: 60px;
   left: 60px;
   width: 315px;
-  margin: 40px 60px 0 0;
+  margin: 40px 0 0 0;
   color: $menu-item;
 }
