+++ dashboard/client/components/Ticket/styles.scss
@@ -1,6 +1,7 @@
 .ticket {
-    background-color: #F9F9F9;
+    background-color: #F3F3F3;
+    border: #E0E1DE solid 1px;
     border-radius: 12px;
-    margin: 3px;
+    margin: 3px 60px 0 60px;
     padding: 10px;
 }
