+++ dashboard/client/pages/profile.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react'
 
-const Profile = () => <>🥰</>
+import Layout from '../components/Layout'
+
+const Profile = () => <Layout>🥰</Layout>
 
 export default Profile
